Re-arrange tests and code location
