+++ app/unanimous/store.py
@@ -1,0 +1,17 @@
+"""
+Cache the details of non-words.
+"""
+
+import pathlib
+
+
+def get_config_dir(basepath=None):
+    """
+    Return the directory to store the cache of PyPi packages
+    """
+    if basepath is None:
+        basepath = pathlib.Path.home()
+    path = basepath / ".unanimous"
+    if not path.exists():
+        path.mkdir()
+    return path
